# docs/swagger.yaml
definitions:
  model.Activity:
    properties:
      activeForm:
        type: string
      bid:
        type: string
      collectNum:
        type: integer
      commentNum:
        type: integer
      createdAt:
        type: string
      endTime:
        type: string
      holderType:
        type: string
      ifRegister:
        type: string
      introduce:
        type: string
      isChecking:
        type: string
      likeNum:
        type: integer
      position:
        type: string
      registerMethod:
        type: string
      showImg:
        type: string
      signer:
        type: string
      startTime:
        type: string
      studentID:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  model.Post:
    properties:
      bid:
        type: string
      collectNum:
        type: integer
      commentNum:
        type: integer
      createdAt:
        type: string
      introduce:
        type: string
      likeNum:
        type: integer
      showImg:
        type: string
      studentID:
        type: string
      title:
        type: string
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      collectact:
        type: string
      collectpost:
        type: string
      id:
        type: integer
      likeact:
        type: string
      likepost:
        type: string
      name:
        type: string
      school:
        type: string
      student_id:
        type: string
    type: object
...
  req.LoginReq:
    properties:
      password:
        type: string
      studentid:
        type: string
    type: object
  req.NumReq:
    properties:
      object:
        type: string
      studentid:
        type: string
      target_id:
        type: string
      type:
        type: integer
    type: object
  req.UpdateNameReq:
    properties:
      new_name:
        type: string
      studentid:
        type: string
    type: object
  req.UserAvatarReq:
    properties:
      avatar_url:
        type: string
      studentid:
        type: string
    type: object
  req.UserSearchReq:
    properties:
      keyword:
        type: string
      studentid:
        type: string
    type: object
...
  resp.ImgBedResp:
    properties:
      access_token:
        type: string
      domain_name:
        type: string
    type: object
  resp.ListActivitiesResp:
    properties:
      bid:
        type: string
      collectNum:
        type: integer
      commentNum:
        type: integer
      detailTime:
        properties:
          endTime:
            type: string
          startTime:
            type: string
        type: object
      holderType:
        type: string
      if_register:
        type: string
      introduce:
        type: string
      isCollect:
        type: string
      isLike:
        type: string
      likeNum:
        type: integer
      position:
        type: string
      showImg:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      userInfo:
        properties:
          avatar:
            type: string
          school:
            type: string
          studentid:
            type: string
          username:
            type: string
        type: object
    type: object
  resp.ListPostsResp:
    properties:
      bid:
        type: string
      collectNum:
        type: integer
      commentNum:
        type: integer
      introduce:
        type: string
      isCollect:
        type: string
      isLike:
        type: string
      likeNum:
        type: integer
      showImg:
        items:
          type: string
        type: array
      title:
        type: string
      userInfo:
        properties:
          avatar:
            type: string
          school:
            type: string
          studentid:
            type: string
          username:
            type: string
        type: object
    type: object
  resp.LoginResp:
    properties:
      Id:
        type: integer
      avatar:
        type: string
      name:
        type: string
      school:
        type: string
      sid:
        type: string
      token:
        type: string
    type: object
  resp.Resp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: 校灵通 API 文档
  title: EventGlide API
paths:
...
  /user/avatar:
    post:
      description: not finished
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户头像更改
        in: body
        name: userAvatarReq
        required: true
        schema:
          $ref: '#/definitions/req.UserAvatarReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 更新头像
      tags:
      - User
  /user/collect/act:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 加载收藏请求
        in: body
        name: cr
        required: true
        schema:
          $ref: '#/definitions/req.NumReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.ListActivitiesResp'
                  type: array
              type: object
      summary: 加载活动收藏
      tags:
      - User
  /user/collect/post:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 加载收藏请求
        in: body
        name: cr
        required: true
        schema:
          $ref: '#/definitions/req.NumReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.ListPostsResp'
                  type: array
              type: object
      summary: 加载帖子收藏
      tags:
      - User
  /user/info/{id}:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      summary: 获取用户信息
      tags:
      - User
  /user/login:
    post:
      parameters:
      - description: 登录请求
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/req.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.LoginResp'
              type: object
      summary: 登录
      tags:
      - User
  /user/logout:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 登出
      tags:
      - User
  /user/search/act:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 搜索请求
        in: body
        name: ureq
        required: true
        schema:
          $ref: '#/definitions/req.UserSearchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Activity'
                  type: array
              type: object
      summary: 搜索用户活动
      tags:
      - User
  /user/search/post:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 搜索请求
        in: body
        name: ureq
        required: true
        schema:
          $ref: '#/definitions/req.UserSearchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Post'
                  type: array
              type: object
      summary: 搜索用户帖子
      tags:
      - User
  /user/token/qiniu:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.ImgBedResp'
              type: object
      summary: 获取七牛云token
      tags:
      - User
  /user/username:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新用户名
        in: body
        name: unr
        required: true
        schema:
          $ref: '#/definitions/req.UpdateNameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 更新用户名
      tags:
      - User
swagger: "2.0"